move(single + multi)
