+++ examples/DragAndDropAvatar.html
@@ -71,7 +71,7 @@
             configs['frame3D_planars'][2]
           );
 
-          const game = new udvizBrowser.SinglePlayerGamePlanar(
+          const game = new udvizBrowser.ExternalGame.SinglePlanarProcess(
             gameObject3D,
             frame3DPlanar,
             assetManager,
