+++ examples/ZeppelinGame.html
@@ -236,7 +236,7 @@
             configs['frame3D_planars'][0]
           );
 
-          const game = new udvizBrowser.SinglePlayerGamePlanar(
+          const game = new udvizBrowser.ExternalGame.SinglePlanarProcess(
             gameObject3D,
             frame3DPlanar,
             assetManager,
