+++ index.html
@@ -100,8 +100,7 @@
 
       createContainer('SideBarWidget');
       createContainer('Widget');
-      createContainer('SinglePlayerGamePlanar');
-      createContainer('MultiPlayerGamePlanar');
+      createContainer('Game');
       createContainer('Frame3D');
       createContainer('THREEUtil');
       createContainer('GUI');
@@ -247,7 +246,7 @@
 
       addProjectElement(
         'Avatar Game',
-        'SinglePlayerGamePlanar',
+        'Game',
         './examples/AvatarGame.html',
         'This example implements a local mini game illustrating an avatar walking in the city',
         './examples/assets/img/mosaic/AvatarGame.png'
@@ -255,7 +254,7 @@
 
       addProjectElement(
         'Avatar Game with Shader',
-        'SinglePlayerGamePlanar',
+        'Game',
         './examples/AvatarGameShader.html',
         'This example implements a local mini game illustrating an avatar walking in the city with collision and shader pass.',
         './examples/assets/img/mosaic/AvatarGameShader.png'
@@ -263,7 +262,7 @@
 
       addProjectElement(
         'Zeppelin Game',
-        'SinglePlayerGamePlanar',
+        'Game',
         './examples/ZeppelinGame.html',
         'This example implements a single player game planar',
         './examples/assets/img/mosaic/ZeppelinGame.png'
@@ -271,16 +270,16 @@
 
       addProjectElement(
         'Drag and Drop Avatar',
-        'SinglePlayerGamePlanar',
+        'Game',
         './examples/DragAndDropAvatar.html',
         'This example allow to drag and drop an avatar on city',
         './examples/assets/img/mosaic/Drag_and_Drop_Avatar.png'
       );
 
       addProjectElement(
-        'Note Game',
-        'MultiPlayerGamePlanar',
-        './examples/MultiPlayerGamePlanar.html',
+        'Multi Player Note Game',
+        'Game',
+        './examples/MultiPlanarProcess.html',
         'This example allow to share note in city',
         './examples/assets/img/mosaic/Note_Game.png'
       );
