+++ packages/browser/src/Component/Game/External/ExternalGame.js
@@ -1,1 +1,3 @@
 export { Context, ExternalScriptBase as ScriptBase } from './Context';
+export { MultiPlanarProcess } from './MultiPlanarProcess';
+export { SinglePlanarProcess } from './SinglePlanarProcess';
